Refactor youtube upload according to comments from https://github.com/edgi-govdata-archiving/edgi-scripts/pull/55
