+++ scripts/auth.py
@@ -10,6 +10,17 @@ from oauth2client.file import Storage
 
 # If modifying these scopes, delete your previously saved credentials
 SCOPES = ['https://www.googleapis.com/auth/youtube.upload', 'https://www.googleapis.com/auth/youtube.force-ssl']
+
+# The CLIENT_SECRET_FILE variable specifies the name of a file that contains
+# the OAuth 2.0 information for this application, including its client_id and
+# client_secret. You can acquire an OAuth 2.0 client ID and client secret from
+# the {{ Google Cloud Console }} at
+# {{ https://cloud.google.com/console }}.
+# Please ensure that you have enabled the YouTube Data API for your project.
+# For more information about using OAuth2 to access the YouTube Data API, see:
+#   https://developers.google.com/youtube/v3/guides/authentication
+# For more information about the client_secrets.json file format, see:
+#   https://developers.google.com/api-client-library/python/guide/aaa_client_secrets
 CLIENT_SECRET_FILE = 'client_secret.json'
 APPLICATION_NAME = 'youtube-upload'
 CREDS_FILENAME = '.youtube-upload-credentials.json'
